+++ lib/click_house/extend/connection_selective.rb
@@ -6,17 +6,17 @@ module ClickHouse
       # @return [ResultSet]
       def select_all(sql)
         response = get(body: sql, query: { default_format: 'JSON' })
-        Response::Factory[response]
+        Response::Factory.response(response, config)
       end
 
       def select_value(sql)
         response = get(body: sql, query: { default_format: 'JSON' })
-        Array(Response::Factory[response].first).dig(0, -1)
+        Array(Response::Factory.response(response, config).first).dig(0, -1)
       end
 
       def select_one(sql)
         response = get(body: sql, query: { default_format: 'JSON' })
-        Response::Factory[response].first
+        Response::Factory.response(response, config).first
       end
     end
   end