+++ lib/click_house/response/factory.rb
@@ -3,19 +3,28 @@
 module ClickHouse
   module Response
     class Factory
-      # @return [String, ResultSet]
+      KEY_META = 'meta'
+      KEY_DATA = 'data'
+      KEY_TOTALS = 'totals'
+      KEY_STATISTICS = 'statistics'
+      KEY_ROWS_BEFORE_LIMIT_AT_LEAST = 'rows_before_limit_at_least'
+
+      # @return [Nil], ResultSet]
       # @params faraday [Faraday::Response]
-      def self.[](faraday)
+      # @params config [Config]
+      def self.response(faraday, config)
         body = faraday.body
 
-        return body if !body.is_a?(Hash) || !(body.key?('meta') && body.key?('data'))
+        return body unless body.is_a?(Hash)
+        return body unless body.key?(config.key(KEY_META)) && body.key?(config.key(KEY_DATA))
 
         ResultSet.new(
-          meta: body.fetch('meta'),
-          data: body.fetch('data'),
-          totals: body['totals'],
-          statistics: body['statistics'],
-          rows_before_limit_at_least: body['rows_before_limit_at_least']
+          config: config,
+          meta: body.fetch(config.key(KEY_META)),
+          data: body.fetch(config.key(KEY_DATA)),
+          totals: body[config.key(KEY_TOTALS)],
+          statistics: body[config.key(KEY_STATISTICS)],
+          rows_before_limit_at_least: body[config.key(KEY_ROWS_BEFORE_LIMIT_AT_LEAST)]
         )
       end