+++ spec/oj_helper.rb
@@ -4,6 +4,7 @@ RSpec.configure do |config|
   config.before(:each) do
     ClickHouse.config do |c|
       c.json_parser = ClickHouse::Middleware::ParseJsonOj
+      c.json_serializer = ClickHouse::Serializer::JsonOjSerializer
     end
   end
 end